+++ src/index.js
@@ -17,12 +17,14 @@ import dataAccess from '@adobe/spacecat-shared-data-access';
 import { resolveSecretsName } from '@adobe/spacecat-shared-utils';
 
 import sqs from './support/sqs.js';
+import apex from './apex/handler.js';
 import cwv from './cwv/handler.js';
 import lhs from './lhs/handler.js';
 import notfound from './notfound/handler.js';
 import backlinks from './backlinks/handler.js';
 
 const HANDLERS = {
+  apex,
   cwv,
   'lhs-mobile': lhs,
   'lhs-desktop': lhs,